vendor: update cityfone apn
An introduction to the problem:

I've been using Essential Phone for a while on CityFone's network,
and I've experienced numerous issues from people's calls not coming
through to very low call volume. It got worse to the point where I held
a friend's phone in one hand while I called my phone which was in the
other hand, and the phone did not ring; Instead it went straight to
voicemail. Incredible!

The solution:

This config was obtained after a chat with CityFone's help-desk.
I informed them of the issue and we went over the default APN settings
field-by-field. After the call, I modified the file in
/system/product/etc/apns-config.xml with the new values (similar to
what this PR does), and lo and behold, it worked!

No more low call volumes.
No missed phone calls.
Receiving MMS and text messages work just fine.
Anyways I will be getting a new phone in a couple days, but I just
wanted to get this PR in before I forget about it.

Change-Id: I470ec042915176e5ba031a2345d79831149916cf